+++ include/mat_reader/mat_reader.hpp
@@ -4,7 +4,7 @@
 #include <fstream>
 #include <ranges>
 #include <span>
-Eigen::MatrixXd loadtxt(const std::filesystem::path &file_loc,
+Eigen::MatrixXf loadtxt(const std::filesystem::path &file_loc,
                         const std::string &delim = " ", int skip_rows = 0);
 
 template <typename Derived>
