+++ include/panel_method/source_doublet_single.hpp
@@ -13,16 +13,16 @@ private:
   std::vector<ComputeTask> wakePanelCompTasks;
 
 protected:
-  Eigen::MatrixXd assembleLhs() override;
-  Eigen::VectorXd assembleRhs() override;
-  Eigen::MatrixXd calculatePanelVelocities() override;
+  Eigen::MatrixXf assembleLhs() override;
+  Eigen::VectorXf assembleRhs() override;
+  Eigen::MatrixXf calculatePanelVelocities() override;
 
 public:
-  Eigen::VectorXd sourceStrength;
+  Eigen::VectorXf sourceStrength;
   SourceDoubletSingle(const PanelGeometry<SurfacePanel> &surfacePanelGeo,
                       const PanelGeometry<WakePanel> &wakePanelGeo,
-                      const EvalPoints<double> &evalPoints,
-                      std::unique_ptr<ISolver> &&solver, double AoAd);
+                      const EvalPoints<float> &evalPoints,
+                      std::unique_ptr<ISolver> &&solver, float AoAd);
   void run() override;
-  Eigen::ArrayXd getSource() const;
+  Eigen::ArrayXf getSource() const;
 };
