+++ include/singularity/__const_source_far.hpp
@@ -8,18 +8,18 @@
 
 struct SourceFar : IConstant3dSingularity<SourceFar> {
 
-  static Eigen::ArrayXd calcInfluenceImpl(const ComputeTask &compTask) {
-    const Eigen::ArrayXd norms = (compTask.points.transpose() -
+  static Eigen::ArrayXf calcInfluenceImpl(const ComputeTask &compTask) {
+    const Eigen::ArrayXf norms = (compTask.points.transpose() -
                                   compTask.face.centrePoint.replicate(
                                       1, compTask.points.transpose().cols()))
                                      .matrix()
                                      .colwise()
                                      .norm();
     return compTask.face.area /
-           (4 * std::numbers::pi_v<double> * norms).array();
+           (4 * std::numbers::pi_v<float> * norms).array();
   }
 
-  static double calcSelfInfluenceImpl(const ComputeTask &compTask) {
+  static float calcSelfInfluenceImpl(const ComputeTask &compTask) {
     return calcInfluenceImpl(compTask)(0);
   }
 };
