+++ include/singularity/const_doublet_far.hpp
@@ -7,6 +7,6 @@
 
 struct DoubletFar : IConstant3dSingularity<DoubletFar> {
 
-  static Eigen::ArrayXd calcInfluenceImpl(const ComputeTask &compTask); 
-  static double calcSelfInfluenceImpl(const ComputeTask &compTask);
+  static Eigen::ArrayXf calcInfluenceImpl(const ComputeTask &compTask); 
+  static float calcSelfInfluenceImpl(const ComputeTask &compTask);
 };
